3.2.1-pre1: 修改zco.py为request.py，增加session，分离文件与日志处理，删除注册表操作
