+++ zco/__init__.py
@@ -1,5 +1,5 @@
-from .zco import *
+from zco.request import *
 
-__version__ = '0.0.1'
+__version__ = '0.0.2'
 __title__ = 'zco'
-__description__ = '提供更简单的网络服务'
+__description__ = '提供更简单的网络请求服务'
