+++ zco/request.py
@@ -6,7 +6,7 @@ import subprocess
 import os
 import time
 
-from .cookies import Cookies
+from zco.cookies import Cookies
 
 retry_times = 0
 browser_list = ['chrome.exe', 'msedge.exe']
@@ -224,6 +224,7 @@ def connect(url, headers=None):
 def exec_cmd(cmd):
     return subprocess.Popen(cmd, shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
 
+
 def open_pixiv(path):
     if not path:
         logging.warning('路径未填写')
@@ -235,4 +236,3 @@ def open_pixiv(path):
     exec_cmd(f'taskkill /F /IM {browser}').communicate()
     pram = '--start-url https://www.pixiv.net --host-rules="MAP api.fanbox.cc api.fanbox.cc,MAP *pixiv.net pixivision.net,MAP *fanbox.cc pixivision.net,MAP *pximg.net U4" --host-resolver-rules="MAP api.fanbox.cc 172.64.146.116,MAP pixivision.net 210.140.139.155,MAP U4 210.140.139.133" --test-type --ignore-certificate-errors'
     return exec_cmd(f'"{path}" {pram}')
-
